requirements.txt
asgiref==3.5.2
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.7
sqlparse==0.4.3
